add common parent class for pmdarima
