+++ sktime/forecasting/arima.py
@@ -5,17 +5,13 @@
 __author__ = ["Markus Löning", "Hongyi Yang"]
 __all__ = ["AutoARIMA", "ARIMA"]
 
-import pandas as pd  # type: ignore
-
-from sktime.forecasting.base._base import DEFAULT_ALPHA
-from sktime.forecasting.base._sktime import _SktimeForecaster
-from sktime.forecasting.base._sktime import _OptionalForecastingHorizonMixin
+from sktime.forecasting.base.adapters._pmdarima import _PmdArimaAdapter
 from sktime.utils.validation._dependencies import _check_soft_dependencies
 
 _check_soft_dependencies("pmdarima")
 
 
-class AutoARIMA(_OptionalForecastingHorizonMixin, _SktimeForecaster):
+class AutoARIMA(_PmdArimaAdapter):
     """Automatically discover the optimal order for an ARIMA model.
 
     The auto-ARIMA process seeks to identify the most optimal parameters
@@ -301,25 +297,11 @@ class AutoARIMA(_OptionalForecastingHorizonMixin, _SktimeForecaster):
 
         super(AutoARIMA, self).__init__()
 
-    def fit(self, y, X=None, fh=None, **fit_params):
-        """Fit to training data.
-
-        Parameters
-        ----------
-        y : pd.Series
-            Target time series to which to fit the forecaster.
-        fh : int, list or np.array, optional (default=None)
-            The forecasters horizon with the steps ahead to to predict.
-        X : pd.DataFrame, optional (default=None)
-            Exogenous variables are ignored
-        Returns
-        -------
-        self : returns an instance of self.
-        """
+    def _instantiate_model(self):
         # import inside method to avoid hard dependency
         from pmdarima.arima import AutoARIMA as _AutoARIMA  # type: ignore
 
-        self._forecaster = _AutoARIMA(
+        return _AutoARIMA(
             start_p=self.start_p,
             d=self.d,
             start_q=self.start_q,
@@ -361,108 +343,8 @@ class AutoARIMA(_OptionalForecastingHorizonMixin, _SktimeForecaster):
             **self.model_kwargs
         )
 
-        self._set_y_X(y, X)
-        self._set_fh(fh)
-        self._forecaster.fit(y, X=X, **fit_params)
-        self._is_fitted = True
-        return self
 
-    def _predict(self, fh, X=None, return_pred_int=False, alpha=DEFAULT_ALPHA):
-        # distinguish between in-sample and out-of-sample prediction
-        fh_oos = fh.to_out_of_sample(self.cutoff)
-        fh_ins = fh.to_in_sample(self.cutoff)
-
-        kwargs = {"X": X, "return_pred_int": return_pred_int, "alpha": alpha}
-
-        # all values are out-of-sample
-        if len(fh_oos) == len(fh):
-            return self._predict_fixed_cutoff(fh_oos, **kwargs)
-
-        # all values are in-sample
-        elif len(fh_ins) == len(fh):
-            return self._predict_in_sample(fh_ins, **kwargs)
-
-        # both in-sample and out-of-sample values
-        else:
-            y_ins = self._predict_in_sample(fh_ins, **kwargs)
-            y_oos = self._predict_fixed_cutoff(fh_oos, **kwargs)
-            return y_ins.append(y_oos)
-
-    def _predict_in_sample(
-        self, fh, X=None, return_pred_int=False, alpha=DEFAULT_ALPHA
-    ):
-        if isinstance(alpha, (list, tuple)):
-            raise NotImplementedError()
-
-        # for in-sample predictions, pmdarima requires zero-based
-        # integer indicies
-        start, end = fh.to_absolute_int(self._y.index[0], self.cutoff)[[0, -1]]
-
-        result = self._forecaster.predict_in_sample(
-            start=start,
-            end=end,
-            X=X,
-            return_conf_int=return_pred_int,
-            alpha=alpha,
-        )
-
-        fh_abs = fh.to_absolute(self.cutoff)
-        fh_idx = fh.to_indexer(self.cutoff, from_cutoff=False)
-
-        if return_pred_int:
-            # unpack and format results
-            y_pred, pred_int = result
-            y_pred = pd.Series(y_pred[fh_idx], index=fh_abs)
-            pred_int = pd.DataFrame(
-                pred_int[fh_idx, :], index=fh_abs, columns=["lower", "upper"]
-            )
-            return y_pred, pred_int
-
-        else:
-            return pd.Series(result[fh_idx], index=fh_abs)
-
-    def _predict_fixed_cutoff(
-        self, fh, X=None, return_pred_int=False, alpha=DEFAULT_ALPHA
-    ):
-        # make prediction
-        n_periods = int(fh.to_relative(self.cutoff)[-1])
-        fh_abs = fh.to_absolute(self.cutoff)
-        fh_idx = fh.to_indexer(self.cutoff)
-
-        result = self._forecaster.predict(
-            n_periods=n_periods,
-            X=X,
-            return_conf_int=return_pred_int,
-            alpha=alpha,
-        )
-
-        if return_pred_int:
-            y_pred, pred_int = result
-            y_pred = pd.Series(y_pred[fh_idx], index=fh_abs)
-            pred_int = pd.DataFrame(
-                pred_int[fh_idx, :], index=fh_abs, columns=["lower", "upper"]
-            )
-            return y_pred, pred_int
-        else:
-            return pd.Series(result[fh_idx], index=fh_abs)
-
-    def get_fitted_params(self):
-        """Get fitted parameters
-
-        Returns
-        -------
-        fitted_params : dict
-        """
-        self.check_is_fitted()
-        names = self._get_fitted_param_names()
-        params = self._forecaster.model_.arima_res_._results.params
-        return {name: param for name, param in zip(names, params)}
-
-    def _get_fitted_param_names(self):
-        return self._forecaster.model_.arima_res_._results.param_names
-
-
-class ARIMA(_OptionalForecastingHorizonMixin, _SktimeForecaster):
+class ARIMA(_PmdArimaAdapter):
     """An ARIMA estimator.
 
     An ARIMA, or autoregressive integrated moving average, is a
@@ -651,25 +533,11 @@ class ARIMA(_OptionalForecastingHorizonMixin, _SktimeForecaster):
 
         super(ARIMA, self).__init__()
 
-    def fit(self, y, X=None, fh=None, **fit_params):
-        """Fit to training data.
-
-        Parameters
-        ----------
-        y : pd.Series
-            Target time series to which to fit the forecaster.
-        fh : int, list or np.array, optional (default=None)
-            The forecasters horizon with the steps ahead to to predict.
-        X : pd.DataFrame, optional (default=None)
-            Exogenous variables are ignored
-        Returns
-        -------
-        self : returns an instance of self.
-        """
+    def _instantiate_model(self):
         # import inside method to avoid hard dependency
         from pmdarima.arima.arima import ARIMA as _ARIMA
 
-        self._forecaster = _ARIMA(
+        return _ARIMA(
             order=self.order,
             seasonal_order=self.seasonal_order,
             start_params=self.start_params,
@@ -683,103 +551,3 @@ class ARIMA(_OptionalForecastingHorizonMixin, _SktimeForecaster):
             with_intercept=self.with_intercept,
             sarimax_kwargs=self.sarimax_kwargs,
         )
-
-        self._set_y_X(y, X)
-        self._set_fh(fh)
-        self._forecaster.fit(y, X=X, **fit_params)
-        self._is_fitted = True
-        return self
-
-    def _predict(self, fh, X=None, return_pred_int=False, alpha=DEFAULT_ALPHA):
-        # distinguish between in-sample and out-of-sample prediction
-        fh_oos = fh.to_out_of_sample(self.cutoff)
-        fh_ins = fh.to_in_sample(self.cutoff)
-
-        kwargs = {"X": X, "return_pred_int": return_pred_int, "alpha": alpha}
-
-        # all values are out-of-sample
-        if len(fh_oos) == len(fh):
-            return self._predict_fixed_cutoff(fh_oos, **kwargs)
-
-        # all values are in-sample
-        elif len(fh_ins) == len(fh):
-            return self._predict_in_sample(fh_ins, **kwargs)
-
-        # both in-sample and out-of-sample values
-        else:
-            y_ins = self._predict_in_sample(fh_ins, **kwargs)
-            y_oos = self._predict_fixed_cutoff(fh_oos, **kwargs)
-            return y_ins.append(y_oos)
-
-    def _predict_in_sample(
-        self, fh, X=None, return_pred_int=False, alpha=DEFAULT_ALPHA
-    ):
-        if isinstance(alpha, (list, tuple)):
-            raise NotImplementedError()
-
-        # for in-sample predictions, pmdarima requires zero-based
-        # integer indicies
-        start, end = fh.to_absolute_int(self._y.index[0], self.cutoff)[[0, -1]]
-
-        result = self._forecaster.predict_in_sample(
-            start=start,
-            end=end,
-            X=X,
-            return_conf_int=return_pred_int,
-            alpha=alpha,
-        )
-
-        fh_abs = fh.to_absolute(self.cutoff)
-        fh_idx = fh.to_indexer(self.cutoff, from_cutoff=False)
-
-        if return_pred_int:
-            # unpack and format results
-            y_pred, pred_int = result
-            y_pred = pd.Series(y_pred[fh_idx], index=fh_abs)
-            pred_int = pd.DataFrame(
-                pred_int[fh_idx, :], index=fh_abs, columns=["lower", "upper"]
-            )
-            return y_pred, pred_int
-
-        else:
-            return pd.Series(result[fh_idx], index=fh_abs)
-
-    def _predict_fixed_cutoff(
-        self, fh, X=None, return_pred_int=False, alpha=DEFAULT_ALPHA
-    ):
-        # make prediction
-        n_periods = int(fh.to_relative(self.cutoff)[-1])
-        fh_abs = fh.to_absolute(self.cutoff)
-        fh_idx = fh.to_indexer(self.cutoff)
-
-        result = self._forecaster.predict(
-            n_periods=n_periods,
-            X=X,
-            return_conf_int=return_pred_int,
-            alpha=alpha,
-        )
-
-        if return_pred_int:
-            y_pred, pred_int = result
-            y_pred = pd.Series(y_pred[fh_idx], index=fh_abs)
-            pred_int = pd.DataFrame(
-                pred_int[fh_idx, :], index=fh_abs, columns=["lower", "upper"]
-            )
-            return y_pred, pred_int
-        else:
-            return pd.Series(result[fh_idx], index=fh_abs)
-
-    def get_fitted_params(self):
-        """Get fitted parameters
-
-        Returns
-        -------
-        fitted_params : dict
-        """
-        self.check_is_fitted()
-        names = self._get_fitted_param_names()
-        params = self._forecaster.arima_res_._results.params
-        return {name: param for name, param in zip(names, params)}
-
-    def _get_fitted_param_names(self):
-        return self._forecaster.arima_res_._results.param_names
